requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
scikit-learn>=1.2.2
pandas>=2.0.0
numpy>=1.24.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=9.5.0

# Model fine-tuning and optimization
peft>=0.4.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
safetensors>=0.3.0

# Experiment tracking
wandb>=0.15.0
tqdm>=4.65.0

# Medical imaging
pydicom>=2.3.0
nibabel>=5.0.0
monai>=1.2.0
opencv-python>=4.7.0
albumentations>=1.3.0

# Fairness and causal inference
fairlearn>=0.8.0
causalml>=0.13.0
dowhy>=0.9.0
causalnex>=0.11.0

# Multi-modal learning
sentencepiece>=0.1.99
tokenizers>=0.13.0
clip>=1.0

# Testing and development
pytest>=7.3.0
jupyter>=1.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Configuration and utilities
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
